Move settings modal
